+++ py/desiutil/bitmask.py
@@ -1,6 +1,9 @@
+# Licensed under a 3-clause BSD style license - see LICENSE.rst
+# -*- coding: utf-8 -*-
 """
-desiutil.maskbits
-=================
+================
+desiutil.bitmask
+================
 
 Mask bits for the spectro pipeline.
 
@@ -14,10 +17,10 @@ Fall 2015
 Example::
 
     #- desispec could create a ccdmask like this
-    
+
     from desiutil.bitmask import BitMask
     import yaml
-    
+
     _bitdefs = yaml.load('''
     ccdmask:
         - [BAD,       0, "Pre-determined bad pixel (any reason)"]
@@ -33,15 +36,15 @@ Example::
 
     #- Accessing the mask
     ccdmask.COSMIC | ccdmask.SATURATED  #- 2**4 + 2**3
-    ccdmask.mask('COSMIC')     #- 2**4, same as ccdmask.COSMIC
-    ccdmask.mask(4)            #- 2**4, same as ccdmask.COSMIC
-    ccdmask.COSMIC             #- 2**4, same as ccdmask.mask('COSMIC')
-    ccdmask.bitnum('COSMIC')   #- 4
-    ccdmask.bitname(4)         #- 'COSMIC'
-    ccdmask.names()            #- ['BAD', 'HOT', 'DEAD', 'SATURATED', 'COSMIC']
-    ccdmask.names(3)           #- ['BAD', 'HOT']
-    ccdmask.comment(0)         #- "Pre-determined bad pixel (any reason)"
-    ccdmask.comment('COSMIC')  #- "Cosmic ray"
+    ccdmask.mask('COSMIC')     # 2**4, same as ccdmask.COSMIC
+    ccdmask.mask(4)            # 2**4, same as ccdmask.COSMIC
+    ccdmask.COSMIC             # 2**4, same as ccdmask.mask('COSMIC')
+    ccdmask.bitnum('COSMIC')   # 4
+    ccdmask.bitname(4)         # 'COSMIC'
+    ccdmask.names()            # ['BAD', 'HOT', 'DEAD', 'SATURATED', 'COSMIC']
+    ccdmask.names(3)           # ['BAD', 'HOT']
+    ccdmask.comment(0)         # "Pre-determined bad pixel (any reason)"
+    ccdmask.comment('COSMIC')  # "Cosmic ray"
 """
 
 
@@ -59,16 +62,20 @@ class _MaskBit(int):
         self._extra = extra
         for key, value in extra.items():
             if hasattr(self, key):
-                raise ValueError('extra key {} not allowed since int already uses that'.format(key))
+                raise AttributeError(
+                    "Bit {0} extra key '{1}' is already in use by int objects.".format(name, key))
             self.__dict__[key] = value
         return self
 
     def __str__(self):
-        return '{:16s} bit {} mask 0x{:X} - {}'.format(
-            self.name, self.bitnum, self.mask, self.comment)
+        return ('{0.name:16s} bit {0.bitnum} mask 0x{0.mask:X} - ' +
+                '{0.comment}').format(self)
+
+    def __repr__(self):
+        return "_MaskBit('{0.name}', {0.bitnum:d}, '{0.comment}')".format(self)
 
 
-#- Class to provide mask bit utility functions
+#  Class to provide mask bit utility functions
 class BitMask(object):
     """BitMask object.
     """
@@ -116,7 +123,7 @@ class BitMask(object):
     def mask(self, name_or_num):
         """Return mask value, e.g.
 
-        bitmask.mask(3)         #- 2**3
+        bitmask.mask(3)         #  2**3
         bitmask.mask('BLAT')
         bitmask.mask('BLAT|FOO')
         """
@@ -134,7 +141,7 @@ class BitMask(object):
         """
         names = list()
         if mask is None:
-            #- return names in sorted order of bitnum
+            #  return names in sorted order of bitnum
             bitnums = [x for x in self._bits.keys() if isinstance(x, int)]
             for bitnum in sorted(bitnums):
                 names.append(self._bits[bitnum].name)
@@ -145,26 +152,27 @@ class BitMask(object):
                     if bitnum in self._bits.keys():
                         names.append(self._bits[bitnum].name)
                     else:
-                        names.append('UNKNOWN'+str(bitnum))
+                        names.append('UNKNOWN' + str(bitnum))
                 bitnum += 1
 
         return names
 
-    #- Allow access via mask.BITNAME
+    #  Allow access via mask.BITNAME
     def __getattr__(self, name):
         if name in self._bits:
             return self._bits[name]
         else:
-            raise AttributeError('Unknown mask bit name '+name)
+            raise AttributeError('Unknown mask bit name ' + name)
 
-    #- What to print
+    #  What to print
     def __repr__(self):
         result = list()
-        result.append(self._name+':')
-        #- return names in sorted order of bitnum
+        result.append(self._name + ':')
+        #  return names in sorted order of bitnum
         bitnums = [x for x in self._bits.keys() if isinstance(x, int)]
         for bitnum in sorted(bitnums):
             bit = self._bits[bitnum]
+            #- format the line for single bit, with or without extra keys
             line = '  - [{:16s} {:2d}, "{}"'.format(
                 bit.name+',', bit.bitnum, bit.comment)
             if len(bit._extra) > 0: